refactor(code): extracted dao functions to separate files
