+++ apps/knowii/components/layout/user-menu.tsx
@@ -6,7 +6,7 @@ import { useCallback } from 'react';
 import { FaSignOutAlt, FaUser } from 'react-icons/fa';
 import { useTranslations } from 'next-intl';
 import { useUserName } from '@knowii/client';
-import { ACCOUNT_URL } from '@knowii/common';
+import { ACCOUNT_URL, HOME_URL } from '@knowii/common';
 
 // eslint-disable-next-line @typescript-eslint/no-empty-interface
 interface UserMenuProps {}
@@ -19,7 +19,7 @@ export function UserMenu(_props: UserMenuProps) {
   const userName = useUserName();
 
   const signOut = useCallback(async () => {
-    await router.push('/');
+    await router.push(HOME_URL);
     supabaseClient.auth.signOut();
   }, [router, supabaseClient.auth]);
