+++ libs/server/src/lib/dao/communities/create-community.dao.fn.ts
@@ -1,0 +1,39 @@
+import { CreateCommunityInput, createCommunityInputSchema, errorInputValidation, getLogger } from '@knowii/common';
+import { Communities, PrismaClient } from '@prisma/client';
+import { generateErrorMessage } from 'zod-error';
+import { errorMessageOptions } from '@knowii/server';
+
+export async function daoFnCreateCommunity(input: CreateCommunityInput, prismaClient: PrismaClient): Promise<Communities> {
+  const logger = getLogger('communities', daoFnCreateCommunity.name);
+
+  logger.info('Creating a new community: %o', input);
+
+  const inputValidationResult = createCommunityInputSchema.safeParse(input);
+
+  if (!inputValidationResult.success) {
+    const errorMessage = generateErrorMessage(inputValidationResult.error.issues, errorMessageOptions);
+    logger.warn(`${errorInputValidation.description}. Error(s) detected: %s`, errorMessage);
+
+    // TODO introduce specific error types
+    throw new Error(errorMessage);
+  }
+
+  try {
+    // FIXME add current user as owner!
+    const createdCommunity = await prismaClient.communities.create({
+      data: {
+        name: input.name,
+        description: input.description,
+        slug: input.slug,
+      },
+    });
+
+    logger.info('Community created successfully: %o', createdCommunity);
+
+    return createdCommunity;
+  } catch (err: unknown) {
+    logger.error('Failed to create the community: %o', err);
+    // FIXME improve error handling
+    throw err;
+  }
+}
